Fix: cleanup inactive FDs in the consumer polling thread
The data polling thread on the consumer relies on nb_fd and
consumer_quit to determine if it can exit, but the polling thread is
also responsible to close inactive streams and there is a case where it
could exit before it does.

On consumer teardown (consumer_quit == 1 and all streams hanging up), if
a relay becomes unreachable, we flag the streams that talk to this relay
as inactive (CONSUMER_ENDPOINT_INACTIVE) and wakeup the data polling
thread to close them. If that thread is already busy handling the hangup
of the remaining streams, it ends up updating the poll array without all
the inactive streams in it and nb_fd can end up being 0, which makes the
polling thread exit.

So we now track the number of inactive streams in update_poll_array()
and prevent the data polling thread to exit if there are inactive
streams. That way the write on the data_poll_pipe is received by this
thread and it can close the inactive streams properly.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>